requirements_dev.txt
pre-commit>=2.0.0
black>=22.1.0
flake8>=4.0.1
isort>=5.10.1
pytest>=7.0.0
pytest-homeassistant-custom-component>=0.12.0
python-gammu==3.2.4
